+++ tests/test_tasks.py
@@ -2471,6 +2471,37 @@ class TestSinglePlotTask(unittest.TestCase):
         self.assertEqual(self.dataset[0],
                          dict_['properties']['properties']['drawing']['label'])
 
+    def test_default_colormap_gets_set_to_plotter(self):
+        colormap = 'viridis'
+        self.prepare_recipe()
+        self.recipe.datasets[self.dataset[0]].data.data = \
+            np.random.random([5, 5])
+        self.recipe.settings['default_colormap'] = colormap
+        self.plotting_task['type'] = 'SinglePlotter2D'
+        self.task.from_dict(self.plotting_task)
+        self.task.recipe = self.recipe
+        self.task.perform()
+        dict_ = self.task.to_dict()
+        self.assertEqual(colormap,
+                         dict_['properties']['properties']['colormap'])
+
+    def test_default_colormap_does_not_override_task_colormap(self):
+        default_colormap = 'viridis'
+        task_colormap = 'viridis_r'
+        self.prepare_recipe()
+        self.recipe.datasets[self.dataset[0]].data.data = \
+            np.random.random([5, 5])
+        self.recipe.settings['default_colormap'] = default_colormap
+        self.plotting_task['type'] = 'SinglePlotter2D'
+        self.plotting_task['properties'] = \
+            {'properties': {'colormap': task_colormap}}
+        self.task.from_dict(self.plotting_task)
+        self.task.recipe = self.recipe
+        self.task.perform()
+        dict_ = self.task.to_dict()
+        self.assertEqual(task_colormap,
+                         dict_['properties']['properties']['colormap'])
+
 
 class TestMultiPlotTask(unittest.TestCase):
     def setUp(self):